Refactor bot into manageable files
New files for
-listing reservations functions
-reservation calendar class
-reservation class
-sauna system tools functions
